Create 作品重要代码.rar
This project is an important code of FPGA innovation competition.Our project is called FPGA-based Portrait recognition and temperature monitoring Music early warning system. The participation platform is Xlinx Artix-7 series.We want to use FPGA to create a music warning system, which is connected to external devices such as camera, speaker and temperature sensor, and USES camera to carry out video monitoring of a place, and USES temperature sensor to detect the local temperature in real time.The system triggers a music alert when the camera detects an outsider in its field of view, or when a temperature sensor detects a temperature higher than a set standard, and the system alerts us by playing music over speakers.